+++ src/main/java/org/osm2world/core/world/modules/RailwayModule.java
@@ -222,8 +222,8 @@ public class RailwayModule extends ConfigurableWorldModule {
 		}
 
 		@Override
-		public float getWidth() {
-			return (float)groundWidth;
+		public double getWidth() {
+			return groundWidth;
 		}
 
 	}
